+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Fixed
 - workaround multiple screens and fullscreen (macOS)
-- try icon size on Linux
+- tray icon size on Linux
 
 ### Changed
 - update icons and graphic materials
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Korean translations
+- more break and microbreak ideas
+- more settings in Contributor's settings
 
 ## [0.19.1] - 2019-02-17
 ### Fixed
@@ -292,7 +294,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - resume/pause functionality for reminder
 - scripts for creating installers for OS X, Windows, Linux
 
-[Unreleased]: https://github.com/hovancik/stretchly/compare/v0.19.0...HEAD
+[Unreleased]: https://github.com/hovancik/stretchly/compare/v0.19.1...HEAD
+[0.19.0]: https://github.com/hovancik/stretchly/compare/v0.19.0...v0.19.1
 [0.19.0]: https://github.com/hovancik/stretchly/compare/v0.18.0...v0.19.0
 [0.18.0]: https://github.com/hovancik/stretchly/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/hovancik/stretchly/compare/v0.16.0...v0.17.0